Fix emojis blending with background color
As they have an alpha channel, we can use it directly for blending
instead of applying the normal text fragment shader that blends it with
the text color and ends up mixing with background as well (particularly
bad on bright background).

Instead of branching on data in the shader, it and storing an alpha
channel for the complete Atlas, it might be wiser to have a separate
Atlas and rendering phase for these glyphs.